Preferences: Move MainConfigPage to its own module
